refactor(xtask):linux rootfs 添加测试集、构造镜像操作分离到独立的文件
