+++ xtask/Cargo.toml
@@ -12,6 +12,7 @@ build = "build.rs"
 clap = { version = "3.1", features = ["derive"] }
 dircpy = "0.3"
 rand = "0.8"
+lazy_static = "1.4.0"
 rcore-fs = { git = "https://github.com/rcore-os/rcore-fs", rev = "1a3246b" }
 rcore-fs-sfs = { git = "https://github.com/rcore-os/rcore-fs", rev = "1a3246b" }
 rcore-fs-fuse = { git = "https://github.com/rcore-os/rcore-fs", rev = "1a3246b" }
